deploy/importer/requirements.txt
geopandas
numpy
numpy
pandas
psycopg2
python-dotenv
geojson